IGAPP-970: Fix tests
